Fixed folder structure and added comments
